Project structure cleaned up and optimized imports
