+++ src/main/java/by/it_academy/jd2/golubev_107/voting_service/controller/servlet/ArtistServlet.java
@@ -1,7 +1,7 @@
 package by.it_academy.jd2.golubev_107.voting_service.controller.servlet;
 
 import by.it_academy.jd2.golubev_107.voting_service.service.IArtistService;
-import by.it_academy.jd2.golubev_107.voting_service.service.dto.ArtistCreateDto;
+import by.it_academy.jd2.golubev_107.voting_service.service.dto.artist.ArtistCreateDto;
 import by.it_academy.jd2.golubev_107.voting_service.service.impl.ArtistServiceImpl;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.annotation.WebServlet;
