+++ src/main/java/by/it_academy/jd2/golubev_107/voting_service/service/IGenreService.java
@@ -1,10 +1,7 @@
 package by.it_academy.jd2.golubev_107.voting_service.service;
 
-import by.it_academy.jd2.golubev_107.voting_service.service.dto.ArtistCreateDto;
-import by.it_academy.jd2.golubev_107.voting_service.service.dto.ArtistOutDto;
-import by.it_academy.jd2.golubev_107.voting_service.service.dto.ArtistVotingDtoSimple;
-import by.it_academy.jd2.golubev_107.voting_service.service.dto.GenreCreateDto;
-import by.it_academy.jd2.golubev_107.voting_service.service.dto.GenreOutDto;
+import by.it_academy.jd2.golubev_107.voting_service.service.dto.genre.GenreCreateDto;
+import by.it_academy.jd2.golubev_107.voting_service.service.dto.genre.GenreOutDto;
 
 import java.util.List;
