+++ src/main/java/by/it_academy/jd2/golubev_107/voting_service/service/IVoteService.java
@@ -1,9 +1,9 @@
 package by.it_academy.jd2.golubev_107.voting_service.service;
 
 
-import by.it_academy.jd2.golubev_107.voting_service.service.dto.ArtistVotingDtoFull;
-import by.it_academy.jd2.golubev_107.voting_service.service.dto.VoteInptDto;
-import by.it_academy.jd2.golubev_107.voting_service.service.dto.VotesResult;
+import by.it_academy.jd2.golubev_107.voting_service.service.dto.artist.ArtistVotingDtoFull;
+import by.it_academy.jd2.golubev_107.voting_service.service.dto.vote.VoteInptDto;
+import by.it_academy.jd2.golubev_107.voting_service.service.dto.vote.VotesResult;
 
 import java.util.List;
