+++ src/main/java/by/it_academy/jd2/golubev_107/voting_service/service/dto/genre/GenreOutDto.java
@@ -1,4 +1,4 @@
-package by.it_academy.jd2.golubev_107.voting_service.service.dto;
+package by.it_academy.jd2.golubev_107.voting_service.service.dto.genre;
 
 import java.util.Objects;
