+++ src/main/java/by/it_academy/jd2/golubev_107/voting_service/service/dto/vote/VoteInptDto.java
@@ -1,5 +1,6 @@
-package by.it_academy.jd2.golubev_107.voting_service.service.dto;
+package by.it_academy.jd2.golubev_107.voting_service.service.dto.vote;
 
+import by.it_academy.jd2.golubev_107.voting_service.service.dto.artist.ArtistVotingDtoSimple;
 import by.it_academy.jd2.golubev_107.voting_service.storage.entity.Comment;
 
 import java.util.Arrays;
