+++ src/main/java/by/it_academy/jd2/golubev_107/voting_service/service/impl/GenreServiceImpl.java
@@ -1,8 +1,8 @@
 package by.it_academy.jd2.golubev_107.voting_service.service.impl;
 
 import by.it_academy.jd2.golubev_107.voting_service.service.IGenreService;
-import by.it_academy.jd2.golubev_107.voting_service.service.dto.GenreCreateDto;
-import by.it_academy.jd2.golubev_107.voting_service.service.dto.GenreOutDto;
+import by.it_academy.jd2.golubev_107.voting_service.service.dto.genre.GenreCreateDto;
+import by.it_academy.jd2.golubev_107.voting_service.service.dto.genre.GenreOutDto;
 import by.it_academy.jd2.golubev_107.voting_service.storage.IGenreStorage;
 import by.it_academy.jd2.golubev_107.voting_service.storage.entity.Genre;
 import by.it_academy.jd2.golubev_107.voting_service.storage.impl.GenreStorageDbImpl;
